[Refactor] Expose TMA base annotation as a standalone pass
